Adapted installer for LRZ

+++ INSTALL/tools/install_GliomaSolver_on_linux-64-bit.sh
@@ -15,6 +15,14 @@ select yn in "Yes" "No"; do
     esac
 done
 
+echo "Are we on the LRZ-Cluster?"
+select yn in "Yes" "No"; do
+    case $yn in
+        Yes ) LRZ="yes"; break;;         #if this variable is set to "yes" this installer will not have root privileges
+        No ) LRZ="no"; break;;
+    esac
+done
+
 
 echo "--------------------------------------"
 echo ">>> Getting prerequsites and external programms   <<<"
@@ -24,34 +32,77 @@ echo "--------------------------------------"
 
 if [ $BI = "yes" ]          # only for Bayesian Inference
 then
-    sudo apt install gcc gfortran gawk libtool texlive-full pandoc pandoc-citeproc  
-    
+
+    if [ $LRZ = "yes" ]
+    then
+
+        module purge
+        module load lrz tempdir     #obligatory
+        module load spack           #requirement for many libs
+        
+        module load intel-parallel-studio   #requires spack
+        #"intel-parallel-studio: using intel wrappers for mpicc, mpif77, etc"
+        #Intel paralell studio includes: gcc, fortran compiler, 
+
+        #gawk and libtool is allready existent at LRZ
+        module load texlive         #requires spack
+        #pandoc is not available on the LRZ (TODO: workaround)
+
+
+        #Alternativ to intel-parallel-studio:
+        #module load gcc/8           #gcc version can be changed here
+        #module load mpi.intel/2018_gcc
+        #module load caf/gfortran    #requires gcc and mpi.intel       
+
+    else
+        sudo apt install gcc gfortran gawk libtool texlive-full pandoc pandoc-citeproc  
+    fi
+
     #Explaination:    
     #gcc                                    # general prerequsites for building anything
     #gfortran                               # needed for mpich
     #gawk                                   # reliable math calculations with gawk (normal awk is shit)
     #libtool                                # needed for gsl
-    #pandoc pandoc-citeproc texlive-full    # converting our results.md into a pdf
-else
-    sudo apt install gcc gfortran gawk
-fi
+    #pandoc pandoc-citeproc texlive-full    # converting our results.md into a pdf (not available on LRZ)
 
+
+else
+    if [ $LRZ = "yes" ]
+    then
+
+        module purge
+        module load lrz tempdir     #obligatory
+        module load spack           #requirement for many libs
+        
+        module load intel-parallel-studio   #requires spack
+        #"intel-parallel-studio: using intel wrappers for mpicc, mpif77, etc"
+        #Intel paralell studio includes: gcc, fortran compiler, 
+
+    else
+        sudo apt install gcc gfortran gawk
+    fi
+fi
 
 echo "--------------------------------------"
 echo ">>> Downloading externa libraries   <<<"
 echo "--------------------------------------"
 
 cd "${SolverDir}"
+
 #wget tdo.sk/~janka/GliomaSolverHome/libs/lib-linux-64-bit/lib.tgz #TODO: update janas package
-wget https://syncandshare.lrz.de/getlink/fi6xQsW6vgKxt7DEoQZzjnuS/lib.tgz
+wget https://syncandshare.lrz.de/dl/fi6xQsW6vgKxt7DEoQZzjnuS/lib.tgz
 tar -zxf lib.tgz
 rm lib.tgz
+#contains: hypre-2.10.0b.tgz myVTK.tgz tbb40_20120613oss.tgz
 
-wget tdo.sk/~janka/GliomaSolverHome/libs/lib-linux-64-bit/inference_libs.tgz
-tar -zxf inference_libs.tgz
-mv inference_libs/* lib/ && rm -r inference_libs
-rm inference_libs.tgz
-
+if [ $LRZ != "yes" ]    #LRZ modules cover the content of this package
+then
+    wget tdo.sk/~janka/GliomaSolverHome/libs/lib-linux-64-bit/inference_libs.tgz
+    tar -zxf inference_libs.tgz
+    mv inference_libs/* lib/ && rm -r inference_libs
+    rm inference_libs.tgz
+    #contains: gsl-src.tgz  mpich-3.2.1-src.tgz
+fi
 
 echo " "
 echo "--------------------------------------"
@@ -61,54 +112,67 @@ echo "--------------------------------------"
 cd lib
 LIB_BASE=$(pwd)                         # ".../GliomaSolver/lib"
 
-gsl_src=gsl-src
-mpich_src=mpich-3.2.1-src
-tbb=tbb40_20120613oss
+
+echo "--------------------------------------"
+echo " Unpacking vtk:"
+echo "--------------------------------------"
+
+#we use VTK 5 even on the LRZ because there were API changes which were not updated in our MRAG implementation
+#https://vtk.org/Wiki/VTK/VTK_6_Migration/Replacement_of_SetInput
+
 vtk=myVTK
-hypre=hypre-2.10.0b
-
-
-tar -zxf ${mpich_src}.tgz
-tar -zxf ${tbb}.tgz
 tar -zxf ${vtk}.tgz
-tar -zxf ${hypre}.tgz
-
-mkdir -p mpich-install
-rm ${tbb}.tgz
 rm ${vtk}.tgz
-rm ${hypre}.tgz
 
-if [ $BI = "yes" ]          # only for Bayesian Inference 
-then
-    tar -zxf ${gsl_src}.tgz
+if [ $LRZ = "yes" ]     #LRZ compatible installation
+    then
 
-    mkdir -p gsl-install
+    #mpicc is available through intel-paralell-studio
+    module load tbb
+    module load hypre
+
+    if [ $BI = "yes" ]          # only for Bayesian Inference 
+    then
+        module load gsl
+        #pi4u_lite (torc_lite, engine_tmcmc) has till to be used from a external source
+
+    fi
+
+else        #default linux installation
+
+    echo "--------------------------------------"
+    echo " Installing mpich:"
+    echo "--------------------------------------"
+
+    mpich_src=mpich-3.2.1-src   #archive name
+    tar -zxf ${mpich_src}.tgz
+    mkdir -p mpich-install
+
+    cd "${mpich_src}"
+    make clean
+    ./configure --prefix="${LIB_BASE}"/mpich-install
+    make
+    make install
+    export PATH="${LIB_BASE}"/mpich-install/bin:$PATH
+
+    echo "---------------"
+    echo "mpicc is set to"
+    which mpicc
+    echo "---------------"
+    cd "${LIB_BASE}"
+    rm ${mpich_src}.tgz
+    rm -r ${mpich_src}
 fi
-
-
-echo "--------------------------------------"
-echo " Installing mpich:"
-echo "--------------------------------------"
-
-cd "${mpich_src}"
-make clean
-./configure --prefix="${LIB_BASE}"/mpich-install
-make
-make install
-export PATH="${LIB_BASE}"/mpich-install/bin:$PATH
-
-echo "---------------"
-echo "mpicc is set to"
-which mpicc
-echo "---------------"
-cd "${LIB_BASE}"
-rm ${mpich_src}.tgz
-rm -r ${mpich_src}
 
 
 echo "--------------------------------------"
 echo " Installing tbb:"
 echo "--------------------------------------"
+#module load tbb #throws error on making that why also on LRZ the prepacked version is used
+
+tbb=tbb40_20120613oss   #archive name
+tar -zxf ${tbb}.tgz
+rm ${tbb}.tgz
 
 cd "${tbb}"
 make clean
@@ -116,31 +180,52 @@ make
 cd "${LIB_BASE}"
 
 
-echo "--------------------------------------"
-echo " Installing Hypre:"
-echo "--------------------------------------"
-cd "${hypre}/src"
-make clean
-./configure
-make
-make install
-cd "${LIB_BASE}"
-
-if [ $BI = "yes" ]          # only for Bayesian Inference 
+if [ $LRZ != "yes" ]
 then
+
+
     echo "--------------------------------------"
-    echo " Installing gsl:"
+    echo " Installing Hypre:"
     echo "--------------------------------------"
 
-    cd "${gsl_src}"
-    ./autogen.sh
-    ./configure --enable-maintainer-mode --disable-dynamic --prefix="${LIB_BASE}"/gsl-install
+    hypre=hypre-2.10.0b     #archive name
+    tar -zxf ${hypre}.tgz
+    rm ${hypre}.tgz
+
+    cd "${hypre}/src"
+    make clean
+    ./configure
     make
     make install
     cd "${LIB_BASE}"
-    rm ${gsl_src}.tgz
-    #rm -r ${gsl_src}   #gonna throw a "rm: remove write-protected regular file 'gsl-src/gsl-config'?" hindering the flow
+fi
 
+if [ $BI = "yes" ]          # only for Bayesian Inference 
+then
+
+    if [ $LRZ != "yes" ]    #default installation
+    then
+        echo "--------------------------------------"
+        echo " Installing gsl:"
+        echo "--------------------------------------"
+
+        gsl_src=gsl-src
+        tar -zxf ${gsl_src}.tgz
+        mkdir -p gsl-install
+
+        cd "${gsl_src}"
+        ./autogen.sh
+        ./configure --enable-maintainer-mode --disable-dynamic --prefix="${LIB_BASE}"/gsl-install
+        make
+        make install
+        cd "${LIB_BASE}"
+        rm ${gsl_src}.tgz
+        #rm -r ${gsl_src}   #gonna throw a "rm: remove write-protected regular file 'gsl-src/gsl-config'?" hindering the flow
+    fi
+
+    #TODO: better way to get torc? maybe clone the current repo from https://github.com/cselab/pi4u ?
+    #We only need the "torc_lite" subdirectory form this repo. However sparse checkjputs are not yet supported properly in git
+    #git clone https://github.com/cselab/pi4u.git
 
     echo "--------------------------------------"
     echo " Installing torc:"
@@ -162,16 +247,28 @@ then
 
     cd "${SolverDir}"/tools/pi4u_lite/Inference
 
-    UserName=$(hostname -s)
     cp tools.make/Makefile .
-    cp tools.make/setup_linux.sh setup_${UserName}.sh
     cp tools.make/run_inference.sh .
-
     sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' Makefile
-    sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' setup_${UserName}.sh
     sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' run_inference.sh
 
-    source setup_${UserName}.sh
+    if [ $LRZ = "yes" ]
+    then
+        cp tools.make/setup_lrz.sh .
+
+        sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' setup_lrz.sh
+
+        source setup_lrz.sh
+
+    else
+        UserName=$(hostname -s)
+        cp tools.make/setup_linux.sh setup_${UserName}.sh
+        
+        sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' setup_${UserName}.sh
+
+        source setup_${UserName}.sh
+    fi
+
     make clean
     make
 
@@ -185,17 +282,31 @@ echo ">>>       Creating Makefile       <<<"
 echo "--------------------------------------"
 cd "${SolverDir}"/makefile
 
-UserName=$(hostname -s)
-cp tools.make/make.linux make.${UserName}
 cp tools.make/Makefile .
-cp tools.make/setup_linux.sh setup_${UserName}.sh
 
-#replace placeholders in files
-sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' make.${UserName}
-sed -i 's|_USER_NAME_|'"${UserName}"'|g' Makefile
-sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' setup_${UserName}.sh
+if [ $LRZ = "yes" ]
+then
+    cp tools.make/make.lrz .
+    cp tools.make/setup_lrz.sh .
 
-source setup_${UserName}.sh
+    #replace placeholders in files
+    sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' make.lrz
+    sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' setup_lrz.sh
+
+    source setup_lrz.sh
+else
+    UserName=$(hostname -s)
+    
+    cp tools.make/make.linux make.${UserName}
+    cp tools.make/setup_linux.sh setup_${UserName}.sh
+
+    #replace placeholders in files
+    sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' make.${UserName}
+    sed -i 's|_USER_NAME_|'"${UserName}"'|g' Makefile
+    sed -i 's|_USER_LIB_BASE_|'"${LIB_BASE}"'|g' setup_${UserName}.sh
+
+    source setup_${UserName}.sh
+fi
 
 echo "  "
 echo "==============================================="
@@ -208,16 +319,27 @@ echo "---------------------------------------"
 
 #cat setup_${UserName}.sh
 echo "LIB_BASE=\"${LIB_BASE}\""
-echo "export LD_LIBRARY_PATH=\$LIB_BASE/tbb40_20120613oss/build/linux_intel64_gcc_cc4.6.1_libc2.5_kernel2.6.18_release/:\$LD_LIBRARY_PATH"
 echo "export LD_LIBRARY_PATH=\$LIB_BASE/myVTK/lib/vtk-5.4/:\$LD_LIBRARY_PATH"
-echo "export LD_LIBRARY_PATH=\$LIB_BASE/hypre-2.10.0b/src/hypre/lib/:\$LD_LIBRARY_PATH"
+echo "export LD_LIBRARY_PATH=\$LIB_BASE/tbb40_20120613oss/build/linux_intel64_gcc_cc4.6.1_libc2.5_kernel2.6.18_release/:\$LD_LIBRARY_PATH"
+    
+if [ $LRZ != "yes" ]
+then
+    echo "export LD_LIBRARY_PATH=\$LIB_BASE/hypre-2.10.0b/src/hypre/lib/:\$LD_LIBRARY_PATH"
+    echo "export PATH=\$LIB_BASE/mpich-install/bin:\$PATH"
+fi
+
 if [ $BI = "yes" ]          # only for Bayesian Inference 
 then
-    echo "export LD_LIBRARY_PATH=\$LIB_BASE/gsl-install/lib/:\$LD_LIBRARY_PATH"
+    if [ $LRZ != "yes" ]
+    then
+        echo "export LD_LIBRARY_PATH=\$LIB_BASE/gsl-install/lib/:\$LD_LIBRARY_PATH"
+    else
+        echo "export LD_LIBRARY_PATH=/lrz/sys/libraries/gsl/2.3/lib:\$LD_LIBRARY_PATH"
+    fi
     echo "export PATH=\$LIB_BASE/usr/torc/bin:\$PATH"
     #cat "${SolverDir}/tools/pi4u_lite/Inference/setup_${UserName}.sh"
 fi
-echo "export PATH=\$LIB_BASE/mpich-install/bin:\$PATH"
+   
 
 echo "---------------------------------------"
 echo "To compile the GliomaSolver do"
